Refractored LoginSignupModal into more composable components
